+++ examples/core/instancing/app.js
@@ -1,4 +1,3 @@
-import 'luma.gl/debug';
 import GL from 'luma.gl/constants';
 import {AnimationLoop, setParameters, pickModels, Cube, picking, dirlight} from 'luma.gl';
 import {Matrix4, radians} from 'math.gl';
