+++ examples/core/transform/app.js
@@ -1,7 +1,6 @@
 /* eslint-disable camelcase */
 /* global window */
-import 'luma.gl/debug';
-import {AnimationLoop, Buffer, setParameters, Model, pickModels, picking} from 'luma.gl';
+import {AnimationLoop, Buffer, Model, pickModels, picking} from 'luma.gl';
 import {_Transform as Transform} from 'luma.gl';
 
 const RED = new Uint8Array([255, 0, 0, 255]);
