+++ website-ocular/src/demos.js
@@ -1,1 +1,28 @@
-export {};
+export {default as Instancing} from '../../examples/core/instancing/app.js';
+export {default as Cubemap} from '../../examples/core/cubemap/app.js';
+// export {default as CustomPicking} from '../../examples/core/custom-picking/app.js';
+// export {default as DeferredRendering} from '../../examples/core/deferred-rendering/app.js';
+export {default as Mandelbrot} from '../../examples/core/mandelbrot/app.js';
+export {default as Fragment} from '../../examples/core/fragment/app.js';
+// export {default as Particles} from '../../examples/core/particles/app.js';
+// export {default as Persistence} from '../../examples/core/persistence/app.js';
+export {default as Picking} from '../../examples/core/picking/app.js';
+export {default as Shadowmap} from '../../examples/core/shadowmap/app.js';
+export {default as TransformFeedback} from '../../examples/core/transform-feedback/app.js';
+
+export {default as Lesson01} from '../../examples/lessons/01/app.js';
+export {default as Lesson02} from '../../examples/lessons/02/app.js';
+export {default as Lesson03} from '../../examples/lessons/03/app.js';
+export {default as Lesson04} from '../../examples/lessons/04/app.js';
+export {default as Lesson05} from '../../examples/lessons/05/app.js';
+export {default as Lesson06} from '../../examples/lessons/06/app.js';
+export {default as Lesson07} from '../../examples/lessons/07/app.js';
+export {default as Lesson08} from '../../examples/lessons/08/app.js';
+export {default as Lesson09} from '../../examples/lessons/09/app.js';
+export {default as Lesson10} from '../../examples/lessons/10/app.js';
+export {default as Lesson11} from '../../examples/lessons/11/app.js';
+export {default as Lesson12} from '../../examples/lessons/12/app.js';
+export {default as Lesson13} from '../../examples/lessons/13/app.js';
+export {default as Lesson16} from '../../examples/lessons/16/app.js';
+
+export default {};
